+++ tools/api-markdown-documenter/src/api-item-transforms/ApiItemTransformUtilities.ts
@@ -356,7 +356,7 @@ export function getAncestralHierarchy(
 /**
  * Determines whether or not the specified API item kind is one that should be rendered to its own document.
  *
- * @remarks This is essentially a wrapper around {@link DocumentationSuiteOptions.documentBoundaries}, but also enforces
+ * @remarks This is essentially a wrapper around {@link DocumentationSuiteConfiguration.documentBoundaries}, but also enforces
  * system-wide invariants.
  *
  * Namely...
@@ -390,7 +390,7 @@ export function doesItemKindRequireOwnDocument(
  *
  * @remarks
  *
- * This is essentially a wrapper around {@link DocumentationSuiteOptions.hierarchyBoundaries}, but also enforces
+ * This is essentially a wrapper around {@link DocumentationSuiteConfiguration.hierarchyBoundaries}, but also enforces
  * system-wide invariants.
  *
  * Namely...
@@ -419,7 +419,7 @@ export function doesItemRequireOwnDocument(
  *
  * @remarks
  *
- * This is essentially a wrapper around {@link DocumentationSuiteOptions.hierarchyBoundaries}, but also enforces
+ * This is essentially a wrapper around {@link DocumentationSuiteConfiguration.hierarchyBoundaries}, but also enforces
  * system-wide invariants.
  *
  * Namely...
@@ -474,7 +474,7 @@ function doesItemGenerateHierarchy(
 /**
  * Determines whether or not the specified API item should have documentation generated for it.
  * This is determined based on its release tag (or inherited release scope) compared to
- * {@link DocumentationSuiteOptions.minimumReleaseLevel}.
+ * {@link DocumentationSuiteConfiguration.minimumReleaseLevel}.
  *
  * @remarks
  *
@@ -538,7 +538,7 @@ export function shouldItemBeIncluded(
 /**
  * Filters and returns the provided list of `ApiItem`s to include only those desired by the user configuration.
  * This is determined based on its release tag (or inherited release scope) compared to
- * {@link DocumentationSuiteOptions.minimumReleaseLevel}.
+ * {@link DocumentationSuiteConfiguration.minimumReleaseLevel}.
  * @param apiItem - The API item being queried.
  * @param config - See {@link ApiItemTransformationConfiguration}.
  *
@@ -554,7 +554,7 @@ export function filterItems(
 /**
  * Filters and returns the child members of the provided `apiItem` to include only those desired by the user configuration.
  * This is determined based on its release tag (or inherited release scope) compared to
- * {@link DocumentationSuiteOptions.minimumReleaseLevel}.
+ * {@link DocumentationSuiteConfiguration.minimumReleaseLevel}.
  * @remarks See {@link shouldItemBeIncluded} for more details.
  * @param apiItem - The API item being queried.
  * @param config - See {@link ApiItemTransformationConfiguration}.