+++ tools/api-markdown-documenter/src/api-item-transforms/default-implementations/TransformApiInterface.ts
@@ -39,7 +39,7 @@ import { createChildDetailsSection, createMemberTables } from "../helpers/index.
  *
  * - index-signatures
  *
- * Details (for any types not rendered to their own documents - see {@link DocumentationSuiteOptions.documentBoundaries})
+ * Details (for any types not rendered to their own documents - see {@link DocumentationSuiteConfiguration.documentBoundaries})
  *
  * - constructor-signatures
  *